fix(sqlgen): bind integral bigint/numeric predicates as exact int64 on both dialects (#281)
Above 2^53 a float64 bind stops being exact, so a bigint predicate could not
address the values #205 legitimately allows in a bigint column:

- parseDuckDBRawParam split the numeric arm — Numeric/BigInt now parse
  int64-first (ToDuckDBParam renders int64 via %d, so MaxInt64 keeps all 19
  digits instead of decimalString's 15-significant-digit %.15g), while
  SmallInt/Integer deliberately stay float64 since their ranges end at 2^31.
- normalizePgEavPayload dropped its float64(v) downcast; value_numeric is
  NUMERIC, so pgx encodes int64 losslessly. This mirrors ConvertPgMainValue,
  which already kept int64 on the main-table path.

Characterization matrix gains a bigint storage class (own file, to keep the
matrix file inside the 500-line limit) covering bound/EAV-only above 2^53 and
MaxInt64. Existing integer and composite rows are re-pinned to the exact int64
EAV args; DuckDB INTEGER binds stay float64 by design.

+++ internal/sqlgen/predicate_normalizer.go
@@ -253,7 +253,9 @@ func normalizePgEavPayload(kv *forma.KvCondition, meta forma.AttributeMetadata, 
 		parsed := numutil.TryParseNumber(valStr)
 		switch v := parsed.(type) {
 		case int64:
-			parsedValue = float64(v)
+			// Exact bind (#281): pgx encodes int64 into the NUMERIC comparison
+			// losslessly. Mirrors ConvertPgMainValue on the main-table path.
+			parsedValue = v
 		case float64:
 			parsedValue = v
 		default:
